started to rewrite everythin

+++ src/com/jfsinternal/BlockIO.java
@@ -1,7 +1,5 @@
 package com.jfsinternal;
 
-import com.jfsmemory.SuperBlkBuffer;
-
 import java.io.File;
 import java.io.FileDescriptor;
 import java.io.IOException;
@@ -18,9 +16,9 @@ public class BlockIO {
     /* File for storing simulated disk's data */
     private static final String DISKFILE = "simdisk.data";
     /* Size of blocks on simulated disk */
-    public static final int BLKSIZE = 128;
+    public static final short BLKSIZE = 128;
     /* Number of blocks on simulated disk */
-    public static final int NUMBLKS = 512;
+    public static final short NUMBLKS = 512;
 
     /* File descriptor of disk data file when open.
      * A null value indicated file is not opened. */
